Deduplicator: Quiet snackbar after deleting individual duplicates
Skip the task-result snackbar when the completed deletion targeted exactly one duplicate file (TargetMode.Duplicates with a single target), since the row already disappears visibly in the list. Shorten the duration of the remaining deletion result snackbars from Long to Short so they do not obscure the next list item for as long.

Adds an isSingleDuplicateDelete extension on SDMTool.Task so both the list and details ViewModels can share the suppression check.

+++ app-tool-deduplicator/src/main/java/eu/darken/sdmse/deduplicator/ui/details/DeduplicatorDetailsScreen.kt
@@ -110,7 +110,7 @@ fun DeduplicatorDetailsScreenHost(
                 is DeduplicatorDetailsViewModel.Event.TaskResult -> snackScope.launch {
                     snackbarHostState.showSnackbar(
                         message = event.result.primaryInfo.get(context),
-                        duration = SnackbarDuration.Long,
+                        duration = SnackbarDuration.Short,
                     )
                 }
